feat: custom hook for dynamic sanity content loading
